requirements.txt
yt-dlp
aiogram
python-dotenv
shazamio
flask